매개변수 타입 변경

+++ back/src/main/java/com/seafood/back/service/ProductService.java
@@ -11,6 +11,6 @@ public interface ProductService {
     public List<ProductEntity> findProductNew();
     public List<OptionEntity> findOption(Integer productId);
     public List<ProductEntity> getProductsByCategoryAndSubcategories(Long categoryId);
-    public List<ProductEntity> getProductsByCategorySub(Integer categoryId);
+    public List<ProductEntity> getProductsByCategorySub(Long categoryId);
     public List<ProductEntity> searchProducts(String query);
 }
